auth

+++ routes/web.php
@@ -16,26 +16,29 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', [App\Http\Controllers\UserController::class,'getAllData'] ) ;
+Route::get('/', [App\Http\Controllers\UserController::class,'getAllData'] )->middleware('auth') ;
 
-Route::get('/job_titles', [App\Http\Controllers\JobTitlesController::class,'getAllJobs'] ) ;
-Route::post('/job_data/save', [App\Http\Controllers\JobTitlesController::class,'SaveJobsData'] ) ;
-Route::get('/job_data_s/edit/{id}', [App\Http\Controllers\JobTitlesController::class,'EditingJobDatas'] ) ;
-Route::post('/job_data_s/updating', [App\Http\Controllers\JobTitlesController::class,'UpdatingJobData'] ) ;
-Route::delete('/job_s_data/delete/{id}', [App\Http\Controllers\JobTitlesController::class,'DeleteJobDatas'] ) ;
+Route::get('/job_titles', [App\Http\Controllers\JobTitlesController::class,'getAllJobs'] )->middleware('auth') ;
+Route::post('/job_data/save', [App\Http\Controllers\JobTitlesController::class,'SaveJobsData'] )->middleware('auth') ;
+Route::get('/job_data_s/edit/{id}', [App\Http\Controllers\JobTitlesController::class,'EditingJobDatas'] )->middleware('auth') ;
+Route::post('/job_data_s/updating', [App\Http\Controllers\JobTitlesController::class,'UpdatingJobData'] )->middleware('auth') ;
+Route::delete('/job_s_data/delete/{id}', [App\Http\Controllers\JobTitlesController::class,'DeleteJobDatas'] )->middleware('auth') ;
 
-Route::get('/depart', [App\Http\Controllers\DepartmentController::class,'getAllDepartment'] ) ;
-Route::post('/departme-nts/save', [App\Http\Controllers\DepartmentController::class,'SaveDepartments'] ) ;
-Route::get('/departme-nts/edit/{id}', [App\Http\Controllers\DepartmentController::class,'EditingDepartment'] ) ;
-Route::post('/departme-nts/edit/updates', [App\Http\Controllers\DepartmentController::class,'UpdateDepartments'] ) ;
-Route::delete('/departme-nts/delete/{id}', [App\Http\Controllers\DepartmentController::class,'DeleteDepartments'] );
+Route::get('/depart', [App\Http\Controllers\DepartmentController::class,'getAllDepartment'] )->middleware('auth') ;
+Route::post('/departme-nts/save', [App\Http\Controllers\DepartmentController::class,'SaveDepartments'] )->middleware('auth') ;
+Route::get('/departme-nts/edit/{id}', [App\Http\Controllers\DepartmentController::class,'EditingDepartment'] )->middleware('auth') ;
+Route::post('/departme-nts/edit/updates', [App\Http\Controllers\DepartmentController::class,'UpdateDepartments'] )->middleware('auth') ;
+Route::delete('/departme-nts/delete/{id}', [App\Http\Controllers\DepartmentController::class,'DeleteDepartments'] )->middleware('auth');
 
-Route::get('/employee_s', [ App\Http\Controllers\EmployeesController::class,'getAllEmployeesDatas']);
-Route::post('/add_employee/data/save/register', [ App\Http\Controllers\EmployeesController::class,'RegisterEmployees']);
-Route::get('/Editing_employee/data/save/Edit/{id}', [ App\Http\Controllers\EmployeesController::class,'EditingEmployeesDatas']);
-Route::post('/updating/edit_employee/save/update', [ App\Http\Controllers\EmployeesController::class,'UpdatingEmployeesDatas']);
-Route::delete('/trash/remove_employee/delete/{id}', [ App\Http\Controllers\EmployeesController::class,'DeleteEmployee']);
+Route::get('/employee_s', [ App\Http\Controllers\EmployeesController::class,'getAllEmployeesDatas'])->middleware('auth');
+Route::post('/add_employee/data/save/register', [ App\Http\Controllers\EmployeesController::class,'RegisterEmployees'])->middleware('auth');
+Route::get('/Editing_employee/data/save/Edit/{id}', [ App\Http\Controllers\EmployeesController::class,'EditingEmployeesDatas'])->middleware('auth');
+Route::post('/updating/edit_employee/save/update', [ App\Http\Controllers\EmployeesController::class,'UpdatingEmployeesDatas'])->middleware('auth');
+Route::delete('/trash/remove_employee/delete/{id}', [ App\Http\Controllers\EmployeesController::class,'DeleteEmployee'])->middleware('auth');
 
+Route::get('/attendence', [ App\Http\Controllers\AttendencesController::class,'getAttendencesEmployees'])->middleware('auth');
+Route::post('/Take_Attandance/Save', [ App\Http\Controllers\AttendencesController::class,'MakeEmployeesAttandances'])->middleware('auth');
+Route::get('/view_Attandance/watch', [ App\Http\Controllers\AttendencesController::class,'ViewAttendence'])->middleware('auth');
 
     //return User::all();
  // return  User::where('email','email')->get();
@@ -43,31 +46,36 @@ Route::delete('/trash/remove_employee/delete/{id}', [ App\Http\Controllers\Emplo
     //return view('welcome');
 //});
 
+Route::get('/home', function () {
+    return redirect('/');
+});
 Route::get('/user', function () {
     return view('user');
-});
+})->middleware('auth');
+
+
 
 Route::get('/job-title', function () {
     return view('job-title');
-});
+})->middleware('auth');
 
 // Route::get('/depart', function () {
 //     return view('depart');
 // });
 Route::get('/icons', function () {
     return view('icons');
-});
+})->middleware('auth');
 
 Route::get('/maps', function () {
     return view('maps');
-});
+})->middleware('auth');
 
 Route::get('/notifications', function () {
     return view('notifications');
-});
+})->middleware('auth');
 
 Route::get('/upgrade-pro', function () {
     return view('upgrade-pro');
-});
+})->middleware('auth');
